Remove failure to delete error in PMs, updated description.

+++ helpful.py
@@ -75,7 +75,8 @@ class Helpful:
         FEU_id = "144670830150811649"
         if ctx.message.server is None or ctx.message.server.id == FEU_id:
             await self.bot.send_message(ctx.message.author, "Your request for moderation was successful.")
-            await self.bot.delete_message(ctx.message)
+            if ctx.message.server is not None:
+                await self.bot.delete_message(ctx.message)
             mod_channel = self.bot.get_channel("185950511567077376")
             paladins = discord.utils.get(ctx.message.server.roles, id="145992793796378624").mention
             await self.bot.send_message(mod_channel, "%s, moderation request received by user %s: Rule %s, at %s." % (paladins, ctx.message.author.name, rule_num, link))
@@ -86,7 +87,7 @@ class Helpful:
     async def howtomod(self):
         """Gives information on how to use the !mod command."""
         await self.bot.say("First, have Developer Mode enabled (Settings -> Appearance -> Developer Mode.")
-        await self.bot.say("Then, click the `...` by the offending message, and click \"Copy ID\".")
+        await self.bot.say("Then, click the `...` by the offending message, and click \"Copy Link\".")
         await self.bot.say("Then simple say !mod <n> <link>, where <n> is the rule it violates, and <link> is the pasted link to the message.")
         await self.bot.say("If you do not have Developer Mode, you may instead of a link, write a short description of where the infraction took place, and by who.")
         await self.bot.say("Note that after requesting moderation, the message requesting moderation will be removed.")
